LayoutTests/http/tests/w3c/dom/nodes/MutationObserver-inner-outer-expected.txt

MutationObservers: innerHTML, outerHTML mutations


FAIL innerHTML mutation assert_equals: mutation records must match expected 2 but got 1
PASS innerHTML with 2 children mutation 
PASS outerHTML mutation 
